Body weight exercises can now be posted to Facebook.

+++ GoMotion/src/com/gomotion/RouteActivity.java
@@ -55,7 +55,7 @@ public class RouteActivity extends Activity
 			
 			map = ((MapFragment) getFragmentManager().findFragmentById(R.id.map)).getMap();
 	
-	        CameraUpdate cam = CameraUpdateFactory.newLatLngZoom(new LatLng(waypoints.getDouble(0), waypoints.getDouble(1)), (float) 15);
+	        CameraUpdate cam = CameraUpdateFactory.newLatLngZoom(new LatLng(waypoints.getDouble(0), waypoints.getDouble(1)), 15);
 	        map.moveCamera(cam);	
 	        
 	        PolylineOptions lineOptions = new PolylineOptions();
